+++ src/RadixDlt.NetworkGateway.PostgresIntegration/Migrations/20241001130214_InitialCreate.cs
@@ -524,6 +524,22 @@ namespace RadixDlt.NetworkGateway.PostgresIntegration.Migrations
                 });
 
             migrationBuilder.CreateTable(
+                name: "key_value_store_totals_history",
+                columns: table => new
+                {
+                    id = table.Column<long>(type: "bigint", nullable: false)
+                        .Annotation("Npgsql:ValueGenerationStrategy", NpgsqlValueGenerationStrategy.IdentityByDefaultColumn),
+                    from_state_version = table.Column<long>(type: "bigint", nullable: false),
+                    entity_id = table.Column<long>(type: "bigint", nullable: false),
+                    total_entries_including_deleted = table.Column<long>(type: "bigint", nullable: false),
+                    total_entries_excluding_deleted = table.Column<long>(type: "bigint", nullable: false)
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("PK_key_value_store_totals_history", x => x.id);
+                });
+
+            migrationBuilder.CreateTable(
                 name: "ledger_transaction_markers",
                 columns: table => new
                 {
@@ -1173,6 +1189,11 @@ namespace RadixDlt.NetworkGateway.PostgresIntegration.Migrations
                 columns: new[] { "key_value_store_entity_id", "from_state_version" });
 
             migrationBuilder.CreateIndex(
+                name: "IX_key_value_store_totals_history_entity_id_from_state_version",
+                table: "key_value_store_totals_history",
+                columns: new[] { "entity_id", "from_state_version" });
+
+            migrationBuilder.CreateIndex(
                 name: "IX_ledger_transaction_markers_entity_id_state_version",
                 table: "ledger_transaction_markers",
                 columns: new[] { "entity_id", "state_version" },
@@ -1485,6 +1506,9 @@ namespace RadixDlt.NetworkGateway.PostgresIntegration.Migrations
 
             migrationBuilder.DropTable(
                 name: "key_value_store_schema_history");
+
+            migrationBuilder.DropTable(
+                name: "key_value_store_totals_history");
 
             migrationBuilder.DropTable(
                 name: "ledger_transaction_markers");